Refactor utils #64
