+++ tests/test_frank_wolfe.py
@@ -3,6 +3,8 @@ import numpy as np
 import pytest
 from scipy import optimize
 import copt as cp
+import copt.constraint
+import copt.loss
 
 np.random.seed(0)
 n_samples, n_features = 20, 16
@@ -14,7 +16,7 @@ b = A.dot(w) + np.random.randn(n_samples)
 # greater than 1
 b = np.abs(b / np.max(np.abs(b)))
 
-LOSS_FUNCS = [cp.utils.LogLoss, cp.utils.SquareLoss]
+LOSS_FUNCS = [copt.loss.LogLoss, copt.loss.SquareLoss]
 
 
 def test_fw_api():
@@ -22,10 +24,10 @@ def test_fw_api():
 
     # test that the algorithm does not fail if x0
     # is a tuple
-    f = cp.utils.LogLoss(A, b, 1.0 / n_samples)
+    f = copt.loss.LogLoss(A, b, 1.0 / n_samples)
     cb = cp.utils.Trace(f)
     alpha = 1.0
-    l1ball = cp.utils.L1Ball(alpha)
+    l1ball = copt.constraint.L1Ball(alpha)
     cp.minimize_frank_wolfe(
         f.f_grad,
         [0] * n_features,
@@ -47,7 +49,7 @@ def test_fw_l1(loss_grad, alpha):
     """Test result of FW algorithm with L1 constraint."""
     f = loss_grad(A, b, 1.0 / n_samples)
     cb = cp.utils.Trace(f)
-    l1ball = cp.utils.L1Ball(alpha)
+    l1ball = copt.constraint.L1Ball(alpha)
     opt = cp.minimize_frank_wolfe(
         f.f_grad,
         np.zeros(n_features),
@@ -93,7 +95,7 @@ def bisection(kw):
 def test_fw_backtrack(obj, step, alpha):
     """Test FW with different options of the line-search strategy."""
     f = obj(A, b, 1.0 / n_samples)
-    traceball = cp.utils.TraceBall(alpha, (4, 4))
+    traceball = copt.constraint.TraceBall(alpha, (4, 4))
     opt = cp.minimize_frank_wolfe(
         f.f_grad,
         np.zeros(n_features),
@@ -118,7 +120,7 @@ def test_pairwise_fw(obj, step, alpha):
     """Test the Pairwise FW method."""
     f = obj(A, b, 1.0 / n_samples)
 
-    l1ball = cp.utils.L1Ball(alpha)
+    l1ball = copt.constraint.L1Ball(alpha)
     x0 = np.zeros(A.shape[1])
     x0[0] = alpha
     cb = cp.utils.Trace(f)
