+++ tests/test_loss.py
@@ -3,6 +3,8 @@ import copt as cp
 from scipy import optimize
 from scipy import sparse
 
+import copt.loss
+
 n_samples, n_features = 100, 10
 A_dense = np.random.randn(n_samples, n_features)
 b = np.random.uniform(0, 1, size=n_samples)
@@ -11,7 +13,7 @@ A_sparse = sparse.rand(n_samples, n_features, density=0.5, format="csr")
 
 def test_loss_grad():
     for A in (A_dense, A_sparse):
-        for loss in [cp.utils.LogLoss, cp.utils.SquareLoss, cp.utils.HuberLoss]:
+        for loss in [copt.loss.LogLoss, copt.loss.SquareLoss, copt.loss.HuberLoss]:
             f = loss(A, b)
             err = optimize.check_grad(
                 f, lambda x: f.f_grad(x)[1], np.random.randn(n_features)
@@ -21,7 +23,7 @@ def test_loss_grad():
 
 def test_log_hess():
     for A in (A_dense, A_sparse):
-        f = cp.utils.LogLoss(A, b)
+        f = copt.loss.LogLoss(A, b)
         x = np.random.randn(n_features)
         Hs = f.hessian_mv(x)
