+++ tests/test_randomized.py
@@ -1,6 +1,8 @@
 import numpy as np
 from scipy import sparse
 import copt as cp
+import copt.loss
+import copt.penalty
 from copt import randomized
 import pytest
 
@@ -25,7 +27,7 @@ all_solvers_unconstrained = (
 
 @pytest.mark.parametrize("name_solver, solver, tol", all_solvers_unconstrained)
 def test_optimize(name_solver, solver, tol):
-    f = cp.utils.LogLoss(A, b)
+    f = copt.loss.LogLoss(A, b)
     for alpha in np.logspace(-3, 3, 3):
         L = cp.utils.get_max_lipschitz(A, "logloss") + alpha / density
         opt = solver(
@@ -38,15 +40,15 @@ def test_optimize(name_solver, solver, tol):
             max_iter=200,
             tol=1e-10,
         )
-        grad = cp.utils.LogLoss(A, b, alpha).f_grad(opt.x)[1]
+        grad = copt.loss.LogLoss(A, b, alpha).f_grad(opt.x)[1]
         assert np.linalg.norm(grad) < tol, name_solver
 
 
 def test_saga_l1():
     alpha = 1.0 / n_samples
-    f = cp.utils.LogLoss(A, b, alpha)
+    f = copt.loss.LogLoss(A, b, alpha)
     for beta in np.logspace(-3, 3, 3):
-        pen = cp.utils.L1Norm(beta)
+        pen = copt.penalty.L1Norm(beta)
         L = cp.utils.get_max_lipschitz(A, "logloss") + alpha / density
 
         for solver in [cp.minimize_saga, cp.minimize_svrg]:
@@ -61,7 +63,7 @@ def test_saga_l1():
                 tol=1e-8,
                 prox=pen.prox_factory(n_features),
             )
-            grad = cp.utils.LogLoss(A, b, alpha).f_grad(opt.x)[1]
+            grad = copt.loss.LogLoss(A, b, alpha).f_grad(opt.x)[1]
             x = opt.x
             ss = 1.0 / L
             # check that the gradient mapping vanishes
@@ -72,7 +74,7 @@ def test_saga_l1():
 def test_vrtos():
     """Test VRTOS with no penalty."""
     alpha = 1.0 / n_samples
-    f = cp.utils.LogLoss(A, b)
+    f = copt.loss.LogLoss(A, b)
     for beta in np.logspace(-3, 3, 3):
         L = cp.utils.get_max_lipschitz(A, "logloss") + alpha / density
 
@@ -86,15 +88,15 @@ def test_vrtos():
             max_iter=200,
         )
 
-        grad = cp.utils.LogLoss(A, b, alpha).f_grad(opt.x)[1]
+        grad = copt.loss.LogLoss(A, b, alpha).f_grad(opt.x)[1]
         assert np.linalg.norm(grad) < 1e-6
 
 
 def test_vrtos_l1():
     alpha = 1.0 / n_samples
-    f = cp.utils.LogLoss(A, b, alpha)
+    f = copt.loss.LogLoss(A, b, alpha)
     for beta in np.logspace(-3, 3, 3):
-        p_1 = cp.utils.L1Norm(beta)
+        p_1 = copt.penalty.L1Norm(beta)
         L = cp.utils.get_max_lipschitz(A, "logloss") + alpha / density
 
         # blocks = np.arange(n_features)
@@ -121,7 +123,7 @@ def test_vrtos_l1():
         )
 
         for x in [opt_1.x, opt_2.x]:
-            full_prox = cp.utils.L1Norm(beta)
+            full_prox = copt.penalty.L1Norm(beta)
             grad = f.f_grad(x)[1]
             ss = 1.0 / L
             # check that the gradient mapping vanishes
@@ -140,9 +142,9 @@ all_groups = [
 @pytest.mark.parametrize("groups", all_groups)
 def test_gl(groups):
     alpha = 1.0 / n_samples
-    f = cp.utils.LogLoss(A, b, alpha)
+    f = copt.loss.LogLoss(A, b, alpha)
     for beta in np.logspace(-3, 3, 3):
-        p_1 = cp.utils.GroupL1(beta, groups)
+        p_1 = copt.penalty.GroupL1(beta, groups)
         L = cp.utils.get_max_lipschitz(A, "logloss") + alpha / density
 
         opt_1 = cp.minimize_vrtos(
@@ -179,7 +181,7 @@ def test_gl(groups):
         )
 
         for x in [opt_1.x, opt_2.x, opt_3.x]:
-            grad = cp.utils.LogLoss(A, b, alpha).f_grad(x)[1]
+            grad = copt.loss.LogLoss(A, b, alpha).f_grad(x)[1]
             ss = 1.0 / L
             # check that the gradient mapping vanishes
             grad_map = (x - p_1.prox(x - ss * grad, ss)) / ss
@@ -191,10 +193,10 @@ def test_vrtos_ogl():
     alpha = 1.0 / n_samples
     groups_1 = [np.arange(8)]
     groups_2 = [np.arange(5, 10)]
-    f = cp.utils.LogLoss(A, b, alpha)
+    f = copt.loss.LogLoss(A, b, alpha)
     for beta in np.logspace(-3, 3, 3):
-        p_1 = cp.utils.GroupL1(beta, groups_1)
-        p_2 = cp.utils.GroupL1(beta, groups_2)
+        p_1 = copt.penalty.GroupL1(beta, groups_1)
+        p_2 = copt.penalty.GroupL1(beta, groups_2)
         L = cp.utils.get_max_lipschitz(A, "logloss") + alpha / density
 
         opt_vrtos = cp.minimize_vrtos(
@@ -224,9 +226,9 @@ def test_vrtos_fl(A_data):
     """Test on overlapping group lasso"""
     n_samples, n_features = A_data.shape
     alpha = 1.0 / n_samples
-    f = cp.utils.LogLoss(A_data, b, alpha)
+    f = copt.loss.LogLoss(A_data, b, alpha)
     for beta in np.logspace(-3, 3, 3):
-        pen = cp.utils.FusedLasso(beta)
+        pen = copt.penalty.FusedLasso(beta)
         L = cp.utils.get_max_lipschitz(A_data, "logloss") + alpha / density
 
         opt_vrtos = cp.minimize_vrtos(
